lession2/code_update/requirements.txt
asgiref==3.7.1
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
mysqlclient==2.1.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.1
tzdata==2023.3